clean out AccessPackageClient
