+++ backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI.Integration/Clients/AccessPackageClient.cs
@@ -55,7 +55,7 @@ namespace Altinn.AccessManagement.UI.Integration.Clients
         /// <inheritdoc />
         public async Task<List<AccessPackage>> GetAccessPackageSearchMatches(string languageCode, string searchString)
         {
-            string endpointUrl = $"package/search/term={searchString}";
+            string endpointUrl = $"meta/info/accesspackages/search/{searchString}";
             string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
 
             HttpResponseMessage response = await _client.GetAsync(token, endpointUrl, languageCode);
@@ -72,110 +72,6 @@ namespace Altinn.AccessManagement.UI.Integration.Clients
 
                 throw error;
             }
-        }
-
-        /// <inheritdoc />
-        public async Task<List<Role>> GetRoleSearchMatches(string languageCode, string searchString)
-        {
-            string endpointUrl = $"role/search/term={searchString}";
-            string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-
-            HttpResponseMessage response = await _client.GetAsync(token, endpointUrl, languageCode);
-
-            return await ClientUtils.DeserializeIfSuccessfullStatusCode<List<Role>>(response, _logger, "AccessPackageClient // GetRoleSearchMatches");
-        }
-
-        /// <inheritdoc />
-        public async Task<List<RoleAssignment>> GetRolesForUser(string languageCode, Guid rightOwnerUuid, Guid rightHolderUuid)
-        {
-            try
-            {
-                var endpointUrl = $"/assignment?from={rightOwnerUuid}&to={rightHolderUuid}";
-                string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-
-                HttpResponseMessage response = await _client.GetAsync(token, endpointUrl, languageCode);
-                string responseContent = await response.Content.ReadAsStringAsync();
-                if (response.StatusCode == HttpStatusCode.OK)
-                {
-                    return JsonSerializer.Deserialize<List<RoleAssignment>>(responseContent, _serializerOptions);
-                }
-                else
-                {
-                    _logger.LogError("AccessManagement.UI // RoleClient // GetRolesForUser // Unexpected HttpStatusCode: {StatusCode}\n {responseBody}", response.StatusCode, responseContent);
-                    return null;
-                }
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "AccessManagement.UI // RoleClient // GetRolesForUser // Exception");
-                return null;
-            }
-        }
-
-        /// <inheritdoc />
-        public async Task<HttpResponseMessage> CreateRoleDelegation(Guid from, Guid to, Guid roleId)
-        {
-             try
-            {
-                var roleAssignment = new RoleAssignment
-                {
-                    RoleId = roleId,
-                    FromId = from,
-                    ToId = to
-                };
-                StringContent requestBody = new StringContent(JsonSerializer.Serialize(roleAssignment, _serializerOptions), Encoding.UTF8, "application/json");
-                var endpointUrl = $"/assignment";
-                string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-                               
-                HttpResponseMessage response = await _client.PostAsync(token, endpointUrl, requestBody);
-                if (response.StatusCode == HttpStatusCode.OK)
-                {
-                    return response;
-                }
-                else
-                {
-                    _logger.LogError("AccessManagement.UI // RegisterClient // GetPartyForOrganization // Unexpected HttpStatusCode: {StatusCode}\n {responseBody}", response.StatusCode, response);
-                    return null;
-                }
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "AccessManagement.UI // RoleClient // GetRolesForUser // Exception");
-                return null;
-            }
-        }
-
-        /// <inheritdoc />
-        public async Task<HttpResponseMessage> DeleteRoleDelegation(Guid assignmentId)
-        {
-             try
-            {
-                var endpointUrl = $"/assignment{assignmentId}";
-                string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-                               
-                HttpResponseMessage response = await _client.DeleteAsync(token, endpointUrl);
-                if (response.StatusCode == HttpStatusCode.OK)
-                {
-                    return response;
-                }
-                else
-                {
-                    _logger.LogError("AccessManagement.UI // RegisterClient // GetPartyForOrganization // Unexpected HttpStatusCode: {StatusCode}\n {responseBody}", response.StatusCode, response);
-                    return null;
-                }
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "AccessManagement.UI // RoleClient // GetRolesForUser // Exception");
-                return null;
-            }
-        }
-
-        /// <inheritdoc />
-        public Task<DelegationCheckResponse> RoleDelegationCheck(Guid rightOwner, Guid roleId)
-        {
-            // TODO: Implement this method when the API is ready
-            throw new NotImplementedException();
         }
     }
 }
