+++ backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI.Mocks/Mocks/AccessPackageClientMock.cs
@@ -39,76 +39,5 @@ namespace Altinn.AccessManagement.UI.Mocks.Mocks
 
             return searchString != null ? Task.FromResult(allPackages.Where(package => package.Name.ToLower().Contains(searchString.ToLower())).ToList()) : Task.FromResult(allPackages);
         }
-
-        public Task<List<Role>> GetRoleSearchMatches(string languageCode, string searchString)
-        {
-            List<Role> allRoles = Util.GetMockData<List<Role>>($"{dataFolder}/Roles/roles.json");
-            return searchString != null ? Task.FromResult(allRoles.Where(role => role.Name.ToLower().Contains(searchString.ToLower())).ToList()) : Task.FromResult(allRoles);
-        }
-
-        /// <inheritdoc />    
-        public Task<List<RoleAssignment>> GetRolesForUser(string languageCode, Guid rightOwnerUuid, Guid rightHolderUuid)
-        {
-            if (rightHolderUuid == Guid.Empty)
-            {
-                throw new Exception("Right holder uuid is not valid");
-            }
-            try
-            {
-                List<RoleAssignment> allAssignments = Util.GetMockData<List<RoleAssignment>>($"{dataFolder}/Roles/GetRolesForUser/{rightHolderUuid}.json");
-                if (allAssignments == null)
-                {
-                    return Task.FromResult(new List<RoleAssignment>());
-                }
-                return Task.FromResult(allAssignments);
-            }
-            catch
-            {
-                return Task.FromResult(new List<RoleAssignment>());
-            }
-        }
-
-        /// <inheritdoc />
-        public Task<HttpResponseMessage> CreateRoleDelegation(Guid from, Guid to, Guid roleId)
-        {
-            if (to == Guid.Empty)
-            {
-                throw new Exception("Right holder uuid is not valid");
-            }
-
-            return Task.FromResult(new HttpResponseMessage(HttpStatusCode.OK));
-        }
-
-        /// <inheritdoc />
-        public Task<HttpResponseMessage> DeleteRoleDelegation(Guid assignmentId)
-        {
-            if (assignmentId == Guid.Empty)
-            {
-                throw new Exception("Right holder uuid is not valid");
-            }
-
-            return Task.FromResult(new HttpResponseMessage(HttpStatusCode.OK));
-        }
-
-        public Task<DelegationCheckResponse> RoleDelegationCheck(Guid rightOwner, Guid roleId)
-        {
-            if (rightOwner == Guid.Empty)
-            {
-                throw new Exception("Right holder uuid is not valid");
-            }
-            try
-            {
-                DelegationCheckResponse res = Util.GetMockData<DelegationCheckResponse>($"{dataFolder}/Roles/DelegationCheck/{roleId}.json");
-                return Task.FromResult(res);
-            }
-            catch
-            {
-                return Task.FromResult(new DelegationCheckResponse()
-                {
-                    CanDelegate = false,
-                    DetailCode = DetailCode.Unknown
-                });
-            }
-        }
     }
 }
