+++ backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI/appsettings.Development.json
@@ -1,7 +1,7 @@
 {
   "PlatformSettings": {
     "ApiAccessManagementEndpoint": "https://platform.at22.altinn.cloud/accessmanagement/api/v1/",
-    "ApiAccessPackageEndpoint": "https://localhost:7065",
+    "ApiAccessPackageEndpoint": "https://platform.at22.altinn.cloud/accessmanagement/api/v1/",
     "ApiAuthenticationEndpoint": "https://platform.at22.altinn.cloud/authentication/api/v1/",
     "ApiProfileEndpoint": "https://platform.at22.altinn.cloud/profile/api/v1/",
     "ApiRegisterEndpoint": "https://platform.at22.altinn.cloud/register/api/v1/",
@@ -15,7 +15,7 @@
   "MockSettings": {
     "AccessManagement_V0": false,
     "AccessManagement": true,
-    "AccessPackage": true,
+    "AccessPackage": false,
     "ResourceRegistry": false,
     "Profile": false,
     "Register": false,
